moved rest.h,cpp and curl_rest.h,cpp to src/ root.
